Abstract Visualizer
Update Visualizer.cs
