+++ Assets/Prefabs/Patrolling/PatrollingVisualizer.prefab
@@ -9,6 +9,8 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 7270184343081284773}
+  - component: {fileID: 449967066000256191}
+  - component: {fileID: 8491385731165973487}
   - component: {fileID: 2357735628227746645}
   m_Layer: 0
   m_Name: PatrollingVisualizer
@@ -32,6 +34,56 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!33 &449967066000256191
+MeshFilter:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1960162275254437419}
+  m_Mesh: {fileID: 0}
+--- !u!23 &8491385731165973487
+MeshRenderer:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1960162275254437419}
+  m_Enabled: 1
+  m_CastShadows: 0
+  m_ReceiveShadows: 1
+  m_DynamicOccludee: 1
+  m_StaticShadowCaster: 0
+  m_MotionVectors: 1
+  m_LightProbeUsage: 1
+  m_ReflectionProbeUsage: 1
+  m_RayTracingMode: 2
+  m_RayTraceProcedural: 0
+  m_RenderingLayerMask: 1
+  m_RendererPriority: 0
+  m_Materials:
+  - {fileID: 2100000, guid: f0e88b9b9b024ee499658a96cd460b98, type: 2}
+  m_StaticBatchInfo:
+    firstSubMesh: 0
+    subMeshCount: 0
+  m_StaticBatchRoot: {fileID: 0}
+  m_ProbeAnchor: {fileID: 0}
+  m_LightProbeVolumeOverride: {fileID: 0}
+  m_ScaleInLightmap: 1
+  m_ReceiveGI: 1
+  m_PreserveUVs: 0
+  m_IgnoreNormalsForChartDetection: 0
+  m_ImportantGI: 0
+  m_StitchLightmapSeams: 1
+  m_SelectedEditorRenderState: 3
+  m_MinimumChartSize: 4
+  m_AutoUVMaxDistance: 0.5
+  m_AutoUVMaxAngle: 89
+  m_LightmapParameters: {fileID: 0}
+  m_SortingLayerID: 0
+  m_SortingLayer: 0
+  m_SortingOrder: 0
+  m_AdditionalVertexStreams: {fileID: 0}
 --- !u!114 &2357735628227746645
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -44,5 +96,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: c36df2aba7ae2ec02873e8d570781ff7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  meshRenderer: {fileID: 8491385731165973487}
+  meshFilter: {fileID: 449967066000256191}
   VertexVisualizer: {fileID: 1182521778900794802, guid: 330b326eb8c9e6de8b1652bfd230f7cc, type: 3}
   EdgeVisualizer: {fileID: 2950328255806474650, guid: 484422f2b87dedfffabe0752d026de4a, type: 3}
